requirements.txt
# requirements.txt
# ClārusAI: AI Literacy Through Cognitive Bias Training
# UCL Master's Dissertation Dependencies

# Core Streamlit Framework
streamlit==1.30.0

# AI Integration - Gemini API
google-generativeai>=0.3.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP & Text Processing
sentence-transformers>=2.2.0
nltk>=3.8.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Statistical Analysis
scipy>=1.11.1
statsmodels>=0.14.0

# Utilities
requests>=2.31.0
jsonschema>=4.17.0

# Live Reload Support
watchdog>=3.0.0

# Development Tools 
black>=23.0.0 # Code formatting
flake8>=6.0.0 # Code linting
pytest>=7.0.0 # Testing framework